Improve design issues in updater.
Revert changes made to method epDbUpdate::_findMatch(), it should not be called
in db level classes like epDbObject. Also move out the logic for finding/renaming
tables from epDbObject, this release the object from keeping a reference to the
manager, as it should be; created a new method for doing the job (_findRelations)
in the updater class. Now, information regards to relationship table is passed to
epDbObject from updater in an array, releasing the object from such unrelated task.
Contribute to a more clean design.