Fix sqlite3 incompatibilities
Tests fixed/added for run over sqlite3. So far there is an important
drawback and is about SQLite3 not supporting any operation for getting
the row number returned from a query result set.

Also add the ability to run DDL queries without returning a result set,
this prevent SQLite3 result set to run the query twice as was the case
when inserting values using the ADODB_sqlite3::_connectionID->query().

For more detail on this look at the api documentation for SQLite3::query
and SQLite3::exec